+++ back-end/src/solicitud/solicitud.service.ts
@@ -4,6 +4,7 @@ import { UpdateSolicitudDto } from './dto/update-solicitud.dto';
 import { SolicitudesRepository } from './repositories/solicitudes.repository';
 import { ISolicitudesService } from './interfaces/solicitudes.service.interface';
 import { Solicitud } from './domain/solicitud.domain';
+import { IllegalArgumentError } from 'src/base/argumentError';
 
 @Injectable()
 export class SolicitudServiceImpl extends ISolicitudesService {
@@ -14,21 +15,28 @@ export class SolicitudServiceImpl extends ISolicitudesService {
     super();
   }
 
-  create(createSolicitudDto: CreateSolicitudDto): Promise<Solicitud> {
-    console.log("dto", createSolicitudDto);
+  async create(createSolicitudDto: CreateSolicitudDto): Promise<Solicitud> {
     const solicitud = new Solicitud(createSolicitudDto);
-    return this.solicitudesRepository.create(solicitud);
+    return await this.solicitudesRepository.create(solicitud);
   }
 
-  findAll(): Promise<Solicitud[]> {
-    return this.solicitudesRepository.getAll();
+  async findAll(): Promise<Solicitud[]> {
+    return await this.solicitudesRepository.getAll();
   }
 
-  findOne(id: string): Promise<Solicitud> {
-    return this.solicitudesRepository.get(id);
+  async findOne(id: string): Promise<Solicitud> {
+    if (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id de la solicitud no puede ser vacio');
+    }
+
+    return await this.solicitudesRepository.get(id);
   }
 
   async update(id: string, updateSolicitudDto: UpdateSolicitudDto) {
+    if (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id de la solicitud no puede ser vacio');
+    }
+
     const solicitud = await this.solicitudesRepository.get(id);
 
     if (solicitud) {
@@ -37,13 +45,16 @@ export class SolicitudServiceImpl extends ISolicitudesService {
         ...updateSolicitudDto,
       });
 
-      await this.solicitudesRepository.update(id, newSolicitud);
-      return newSolicitud;
+      return await this.solicitudesRepository.update(id, newSolicitud);
     }
   }
 
-  remove(id: string): Promise<Solicitud> {
-    return this.solicitudesRepository.delete(id);
+  async remove(id: string): Promise<Solicitud> {
+    if (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id de la solicitud no puede ser vacio');
+    }
+    
+    return await this.solicitudesRepository.delete(id);
   }
 
 }
