+++ back-end/src/users/users.service.ts
@@ -4,6 +4,8 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './domain/user.domain';
 import { UsersRepository } from './repositories/users.repository';
+import { ConflictError } from 'src/base/conflictError';
+import { IllegalArgumentError } from 'src/base/argumentError';
 
 export enum Role {
   User = 'user',
@@ -22,9 +24,8 @@ export class UsersServiceImpl implements IUserService {
     // Comprobar si existe el usuario
     const user = await this.usersRepository.getByName(createUserDto.username);
 
-    console.log(user);
     if (user) {
-      throw new ConflictException('User already exists!');
+      throw new ConflictError('User already exists!');
     }
 
     const newUser = new User({
@@ -32,38 +33,50 @@ export class UsersServiceImpl implements IUserService {
       role: Role.User,
     });
 
-    return this.usersRepository.create(newUser);
+    return await this.usersRepository.create(newUser);
   }
 
-  findAll(): Promise<any[]> {
-    return this.usersRepository.getAll();
+  async findAll(): Promise<User[]> {
+    return await this.usersRepository.getAll();
   }
 
-  findOne(id: string) {
-    return this.usersRepository.get(id);
+  async findOne(id: string) {
+    if (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id del usuario no puede ser vacio');
+    }
+
+    return await this.usersRepository.get(id);
   }
 
-  update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateUserDto: UpdateUserDto) {
+    if (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id del usuario no puede ser vacio');
+    }
 
     //Comprobar si existe el usuario
-    const user = this.usersRepository.get(id);
+    const user = await this.usersRepository.get(id);
 
-    if (!user) {
-      throw new ConflictException('User does not exists!');
-    }
-    
     const updateUser = new User({
+      ...user,
       ...updateUserDto,
     });
 
-    return this.usersRepository.update(id, updateUser);
+    return await this.usersRepository.update(id, updateUser);
   }
 
-  remove(id: string): Promise<any> {
-    return this.usersRepository.delete(id);
+  async remove(id: string): Promise<User> {
+    if  (id === null || id.trim() === '') {
+      throw new IllegalArgumentError('El id del usuario no puede ser vacio');
+    }
+
+    return await this.usersRepository.delete(id);
   }
 
   async getByName(name: string): Promise<User> {
-    return this.usersRepository.getByName(name);
+    if (name === null || name.trim() === '') {
+      throw new IllegalArgumentError('El nombre del usuario no puede ser vacio');
+    }
+
+    return await this.usersRepository.getByName(name);
   }
 }
